rename to NxBinaryDonutChart, use forwardRef, allow all svg attributes
