--- gallery/src/components/NxDonutChart/NxDonutChartExample.tsx
@@ -1,17 +1,0 @@
-/*
- * Copyright (c) 2019-present Sonatype, Inc.
- * This program and the accompanying materials are made available under
- * the terms of the Eclipse Public License 2.0 which accompanies this
- * distribution and is available at https://www.eclipse.org/legal/epl-2.0/.
- */
-import React from 'react';
-import { NxDonutChart } from '@sonatype/react-shared-components';
-
-export default function NxDonutChartExample() {
-  return <NxDonutChart id="donut-chart-id"
-                       percent={90}
-                       outerRadius={30}
-                       innerRadius={10}
-                       fillColors={['#97cbed', '#006bbf']}
-  />;
-}
