+++ lib/src/index.ts
@@ -134,4 +134,4 @@ export {
 export * from './util/threatLevels';
 export * from './util/validationUtil';
 
-export { default as NxDonutChart } from './components/NxDonutChart/NxDonutChart';
+export { default as NxBinaryDonutChart } from './components/NxBinaryDonutChart/NxBinaryDonutChart';
